Module aws_sdk_rdsdata::model

Modules

column_metadata

See ColumnMetadata

record

See Record

result_frame

See ResultFrame

result_set_metadata

See ResultSetMetadata

result_set_options

See ResultSetOptions

sql_parameter

See SqlParameter

sql_statement_result

See SqlStatementResult

struct_value

See StructValue

update_result

See UpdateResult

Structs

ColumnMetadata

Contains the metadata for a column.

Record

A record returned by a call.

ResultFrame

The result set returned by a SQL statement.

ResultSetMetadata

The metadata of the result set returned by a SQL statement.

ResultSetOptions

Options that control how the result set is returned.

SqlParameter

A parameter used in a SQL statement.

SqlStatementResult

The result of a SQL statement.

StructValue

A structure value returned by a call.

UpdateResult

The response elements represent the results of an update.

Enums

ArrayValue

DecimalReturnType

Field

TypeHint

Value